scripts/requirements.txt
anthropic>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
smtplib2>=0.2.1
jinja2>=3.1.0
colorama>=0.4.6
tqdm>=4.66.0
ratelimit>=2.2.1
